LPS-105341 Deduplicating
