+++ modules/apps/layout/layout-seo-web/src/main/java/com/liferay/layout/seo/web/internal/servlet/taglib/ui/BaseLayoutScreenNavigationEntry.java
@@ -1,0 +1,173 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.layout.seo.web.internal.servlet.taglib.ui;
+
+import com.liferay.document.library.kernel.service.DLAppService;
+import com.liferay.document.library.util.DLURLHelper;
+import com.liferay.dynamic.data.mapping.storage.StorageEngine;
+import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
+import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
+import com.liferay.layout.admin.constants.LayoutScreenNavigationEntryConstants;
+import com.liferay.layout.seo.canonical.url.LayoutSEOCanonicalURLProvider;
+import com.liferay.layout.seo.kernel.LayoutSEOLinkManager;
+import com.liferay.layout.seo.service.LayoutSEOSiteLocalService;
+import com.liferay.layout.seo.web.internal.constants.LayoutSEOWebKeys;
+import com.liferay.layout.seo.web.internal.display.context.LayoutsSEODisplayContext;
+import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.model.Group;
+import com.liferay.portal.kernel.model.Layout;
+import com.liferay.portal.kernel.model.LayoutConstants;
+import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.service.GroupLocalService;
+import com.liferay.portal.kernel.service.LayoutLocalService;
+import com.liferay.portal.kernel.util.AggregateResourceBundle;
+import com.liferay.portal.kernel.util.JavaConstants;
+import com.liferay.portal.kernel.util.Portal;
+import com.liferay.portal.kernel.util.ResourceBundleUtil;
+
+import java.io.IOException;
+
+import java.util.Locale;
+import java.util.Objects;
+import java.util.ResourceBundle;
+
+import javax.portlet.PortletRequest;
+import javax.portlet.RenderResponse;
+
+import javax.servlet.ServletContext;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Alejandro Tardín
+ */
+public abstract class BaseLayoutScreenNavigationEntry
+	implements ScreenNavigationEntry<Layout> {
+
+	@Override
+	public String getCategoryKey() {
+		return LayoutScreenNavigationEntryConstants.CATEGORY_KEY_GENERAL;
+	}
+
+	@Override
+	public String getLabel(Locale locale) {
+		return LanguageUtil.get(_getResourceBundle(locale), getEntryKey());
+	}
+
+	@Override
+	public String getScreenNavigationKey() {
+		return LayoutScreenNavigationEntryConstants.
+			SCREEN_NAVIGATION_KEY_LAYOUT;
+	}
+
+	@Override
+	public boolean isVisible(User user, Layout layout) {
+		Group group = groupLocalService.fetchGroup(layout.getGroupId());
+
+		if ((group != null) && group.isLayoutPrototype()) {
+			return false;
+		}
+
+		Layout draftLayout = layoutLocalService.fetchLayout(
+			portal.getClassNameId(Layout.class), layout.getPlid());
+
+		if ((Objects.equals(
+				layout.getType(), LayoutConstants.TYPE_ASSET_DISPLAY) ||
+			 Objects.equals(layout.getType(), LayoutConstants.TYPE_CONTENT)) &&
+			(draftLayout == null)) {
+
+			return false;
+		}
+
+		return true;
+	}
+
+	@Override
+	public void render(
+			HttpServletRequest httpServletRequest,
+			HttpServletResponse httpServletResponse)
+		throws IOException {
+
+		httpServletRequest.setAttribute(
+			LayoutSEOWebKeys.LAYOUT_PAGE_LAYOUT_SEO_DISPLAY_CONTEXT,
+			new LayoutsSEODisplayContext(
+				dlAppService, dlurlHelper, layoutSEOCanonicalURLProvider,
+				layoutSEOLinkManager, layoutSEOSiteLocalService,
+				portal.getLiferayPortletRequest(
+					(PortletRequest)httpServletRequest.getAttribute(
+						JavaConstants.JAVAX_PORTLET_REQUEST)),
+				portal.getLiferayPortletResponse(
+					(RenderResponse)httpServletRequest.getAttribute(
+						JavaConstants.JAVAX_PORTLET_RESPONSE)),
+				storageEngine));
+
+		jspRenderer.renderJSP(
+			_servletContext, httpServletRequest, httpServletResponse,
+			getJspPath());
+	}
+
+	@Reference(
+		target = "(osgi.web.symbolicname=com.liferay.layout.seo.web)",
+		unbind = "-"
+	)
+	public void setServletContext(ServletContext servletContext) {
+		_servletContext = servletContext;
+	}
+
+	protected abstract String getJspPath();
+
+	@Reference
+	protected DLAppService dlAppService;
+
+	@Reference
+	protected DLURLHelper dlurlHelper;
+
+	@Reference
+	protected GroupLocalService groupLocalService;
+
+	@Reference
+	protected JSPRenderer jspRenderer;
+
+	@Reference
+	protected LayoutLocalService layoutLocalService;
+
+	@Reference
+	protected LayoutSEOCanonicalURLProvider layoutSEOCanonicalURLProvider;
+
+	@Reference
+	protected LayoutSEOLinkManager layoutSEOLinkManager;
+
+	@Reference
+	protected LayoutSEOSiteLocalService layoutSEOSiteLocalService;
+
+	@Reference
+	protected Portal portal;
+
+	@Reference
+	protected StorageEngine storageEngine;
+
+	private ResourceBundle _getResourceBundle(Locale locale) {
+		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
+			"content.Language", locale, getClass());
+
+		return new AggregateResourceBundle(
+			resourceBundle, portal.getResourceBundle(locale));
+	}
+
+	private ServletContext _servletContext;
+
+}
